+++ xds/src/main/java/io/grpc/xds/XdsLoadBalancer.java
@@ -32,9 +32,8 @@ import io.grpc.NameResolver.ConfigOrError;
 import io.grpc.Status;
 import io.grpc.SynchronizationContext.ScheduledHandle;
 import io.grpc.internal.ServiceConfigUtil.LbConfig;
+import io.grpc.xds.SubchannelStore.SubchannelStoreImpl;
 import io.grpc.xds.XdsComms.AdsStreamCallback;
-import io.grpc.xds.XdsLbState.SubchannelStore;
-import io.grpc.xds.XdsLbState.SubchannelStoreImpl;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
