+++ xds/src/test/java/io/grpc/xds/XdsLbStateTest.java
@@ -45,13 +45,12 @@ import io.grpc.SynchronizationContext;
 import io.grpc.internal.FakeClock;
 import io.grpc.testing.GrpcCleanupRule;
 import io.grpc.xds.InterLocalityPicker.WeightedChildPicker;
+import io.grpc.xds.SubchannelStore.SubchannelStoreImpl;
+import io.grpc.xds.SubchannelStore.SubchannelStoreImpl.PickerFactory;
 import io.grpc.xds.XdsComms.AdsStreamCallback;
 import io.grpc.xds.XdsLbState.LbEndpoint;
 import io.grpc.xds.XdsLbState.Locality;
 import io.grpc.xds.XdsLbState.LocalityInfo;
-import io.grpc.xds.XdsLbState.SubchannelStore;
-import io.grpc.xds.XdsLbState.SubchannelStoreImpl;
-import io.grpc.xds.XdsLbState.SubchannelStoreImpl.PickerFactory;
 import java.net.InetSocketAddress;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
